Added Logout leaving page and custom logout flow. Improves UX.
