+++ admin/src/pages/LeavingPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import LeavingIcon from 'material-ui/svg-icons/maps/flight';
+
+import WaitingPage from './WaitingPage';
+
+const LeavingPage = () => {
+	return <WaitingPage icon={LeavingIcon} />
+}
+
+export default LeavingPage;
